# Add support for attributes with array values