--- a/tempodb/encoding/vparquet3/block_traceql.go
+++ b/tempodb/encoding/vparquet3/block_traceql.go
@@ -167,7 +167,7 @@ func (s *span) DurationNanos() uint64 {
 	return s.durationNanos
 }
 
-func (s *span) DescendantOf(lhs []traceql.Span, rhs []traceql.Span, falseForAll bool, invert bool, buffer []traceql.Span) []traceql.Span {
+func (s *span) DescendantOf(lhs, rhs []traceql.Span, falseForAll, invert bool, buffer []traceql.Span) []traceql.Span {
 	if len(lhs) == 0 || len(rhs) == 0 {
 		return nil
 	}
@@ -183,7 +183,7 @@ func (s *span) DescendantOf(lhs []traceql.Span, rhs []traceql.Span, falseForAll 
 	}
 	sort.Slice(lhs, sortFn)
 
-	descendantOf := func(a *span, b *span) bool {
+	descendantOf := func(a, b *span) bool {
 		if a.nestedSetLeft == 0 ||
 			b.nestedSetLeft == 0 ||
 			a.nestedSetRight == 0 ||
@@ -229,13 +229,13 @@ func (s *span) DescendantOf(lhs []traceql.Span, rhs []traceql.Span, falseForAll 
 	return buffer
 }
 
-func (s *span) SiblingOf(lhs []traceql.Span, rhs []traceql.Span, falseForAll bool, buffer []traceql.Span) []traceql.Span {
+func (s *span) SiblingOf(lhs, rhs []traceql.Span, falseForAll bool, buffer []traceql.Span) []traceql.Span {
 	// this is easy. we're just looking for anything on the lhs side with the same nested set parent as the rhs
 	sort.Slice(lhs, func(i, j int) bool {
 		return lhs[i].(*span).nestedSetParent < lhs[j].(*span).nestedSetParent
 	})
 
-	siblingOf := func(a *span, b *span) bool {
+	siblingOf := func(a, b *span) bool {
 		return a.nestedSetParent == b.nestedSetParent &&
 			a.nestedSetParent != 0 &&
 			b.nestedSetParent != 0
@@ -272,7 +272,7 @@ func (s *span) SiblingOf(lhs []traceql.Span, rhs []traceql.Span, falseForAll boo
 	return buffer
 }
 
-func (s *span) ChildOf(lhs []traceql.Span, rhs []traceql.Span, falseForAll bool, invert bool, buffer []traceql.Span) []traceql.Span {
+func (s *span) ChildOf(lhs, rhs []traceql.Span, falseForAll, invert bool, buffer []traceql.Span) []traceql.Span {
 	// we will search the LHS by either nestedSetLeft or nestedSetParent. if we are doing child we sort by nestedSetLeft
 	// so we can quickly find children. if the invert flag is set we are looking for parents and so we sort appropriately
 	sortFn := func(i, j int) bool { return lhs[i].(*span).nestedSetLeft < lhs[j].(*span).nestedSetLeft }
@@ -280,7 +280,7 @@ func (s *span) ChildOf(lhs []traceql.Span, rhs []traceql.Span, falseForAll bool,
 		sortFn = func(i, j int) bool { return lhs[i].(*span).nestedSetParent < lhs[j].(*span).nestedSetParent }
 	}
 
-	childOf := func(a *span, b *span) bool {
+	childOf := func(a, b *span) bool {
 		return a.nestedSetLeft == b.nestedSetParent &&
 			a.nestedSetLeft != 0 &&
 			b.nestedSetParent != 0
@@ -448,10 +448,10 @@ const (
 	columnPathServiceStatsSpanCount    = "ServiceStats.key_value.value.SpanCount"
 	columnPathServiceStatsErrorCount   = "ServiceStats.key_value.value.ErrorCount"
 	columnPathResourceAttrKey          = "rs.list.element.Resource.Attrs.list.element.Key"
-	columnPathResourceAttrString       = "rs.list.element.Resource.Attrs.list.element.Value"
-	columnPathResourceAttrInt          = "rs.list.element.Resource.Attrs.list.element.ValueInt"
-	columnPathResourceAttrDouble       = "rs.list.element.Resource.Attrs.list.element.ValueDouble"
-	columnPathResourceAttrBool         = "rs.list.element.Resource.Attrs.list.element.ValueBool"
+	columnPathResourceAttrString       = "rs.list.element.Resource.Attrs.list.element.Value.list.element"
+	columnPathResourceAttrInt          = "rs.list.element.Resource.Attrs.list.element.ValueInt.list.element"
+	columnPathResourceAttrDouble       = "rs.list.element.Resource.Attrs.list.element.ValueDouble.list.element"
+	columnPathResourceAttrBool         = "rs.list.element.Resource.Attrs.list.element.ValueBool.list.element"
 	columnPathResourceServiceName      = "rs.list.element.Resource.ServiceName"
 	columnPathResourceCluster          = "rs.list.element.Resource.Cluster"
 	columnPathResourceNamespace        = "rs.list.element.Resource.Namespace"
@@ -470,10 +470,10 @@ const (
 	columnPathSpanStatusCode     = "rs.list.element.ss.list.element.Spans.list.element.StatusCode"
 	columnPathSpanStatusMessage  = "rs.list.element.ss.list.element.Spans.list.element.StatusMessage"
 	columnPathSpanAttrKey        = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.Key"
-	columnPathSpanAttrString     = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.Value"
-	columnPathSpanAttrInt        = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueInt"
-	columnPathSpanAttrDouble     = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueDouble"
-	columnPathSpanAttrBool       = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueBool"
+	columnPathSpanAttrString     = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.Value.list.element"
+	columnPathSpanAttrInt        = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueInt.list.element"
+	columnPathSpanAttrDouble     = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueDouble.list.element"
+	columnPathSpanAttrBool       = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueBool.list.element"
 	columnPathSpanHTTPStatusCode = "rs.list.element.ss.list.element.Spans.list.element.HttpStatusCode"
 	columnPathSpanHTTPMethod     = "rs.list.element.ss.list.element.Spans.list.element.HttpMethod"
 	columnPathSpanHTTPURL        = "rs.list.element.ss.list.element.Spans.list.element.HttpUrl"