--- a/tempodb/encoding/vparquet3/schema.go
+++ b/tempodb/encoding/vparquet3/schema.go
@@ -49,18 +49,33 @@ const (
 	DefinitionLevelResourceAttrs             = 2
 	DefinitionLevelResourceSpansILSSpan      = 3
 	DefinitionLevelResourceSpansILSSpanAttrs = 4
+	// TODO do we need a new definition level for attr values?
 
 	FieldResourceAttrKey       = "rs.list.element.Resource.Attrs.list.element.Key"
-	FieldResourceAttrVal       = "rs.list.element.Resource.Attrs.list.element.Value"
-	FieldResourceAttrValInt    = "rs.list.element.Resource.Attrs.list.element.ValueInt"
-	FieldResourceAttrValDouble = "rs.list.element.Resource.Attrs.list.element.ValueDouble"
-	FieldResourceAttrValBool   = "rs.list.element.Resource.Attrs.list.element.ValueBool"
+	FieldResourceAttrVal       = "rs.list.element.Resource.Attrs.list.element.Value.list.element"
+	FieldResourceAttrValInt    = "rs.list.element.Resource.Attrs.list.element.ValueInt.list.element"
+	FieldResourceAttrValDouble = "rs.list.element.Resource.Attrs.list.element.ValueDouble.list.element"
+	FieldResourceAttrValBool   = "rs.list.element.Resource.Attrs.list.element.ValueBool.list.element"
 
 	FieldSpanAttrKey       = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.Key"
-	FieldSpanAttrVal       = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.Value"
-	FieldSpanAttrValInt    = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueInt"
-	FieldSpanAttrValDouble = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueDouble"
-	FieldSpanAttrValBool   = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueBool"
+	FieldSpanAttrVal       = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.Value.list.element"
+	FieldSpanAttrValInt    = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueInt.list.element"
+	FieldSpanAttrValDouble = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueDouble.list.element"
+	FieldSpanAttrValBool   = "rs.list.element.ss.list.element.Spans.list.element.Attrs.list.element.ValueBool.list.element"
+)
+
+type AttrType int32
+
+const (
+	attrTypeNotSupported AttrType = iota
+	attrTypeString
+	attrTypeInt
+	attrTypeDouble
+	attrTypeBool
+	attrTypeStringArray
+	attrTypeIntArray
+	attrTypeDoubleArray
+	attrTypeBoolArray
 )
 
 var (
@@ -115,11 +130,12 @@ type Attribute struct {
 	Key string `parquet:",snappy,dict"`
 
 	// This is a bad design that leads to millions of null values. How can we fix this?
-	Value        *string  `parquet:",dict,snappy,optional"`
-	ValueInt     *int64   `parquet:",snappy,optional"`
-	ValueDouble  *float64 `parquet:",snappy,optional"`
-	ValueBool    *bool    `parquet:",snappy,optional"`
-	ValueDropped string   `parquet:",snappy,optional"`
+	ValueType    AttrType  `parquet:",snappy"`
+	Value        []string  `parquet:",snappy,dict,list"`
+	ValueInt     []int64   `parquet:",snappy,list"`
+	ValueDouble  []float64 `parquet:",snappy,list"`
+	ValueBool    []bool    `parquet:",snappy,list"`
+	ValueDropped string    `parquet:",snappy,optional"`
 }
 
 // DedicatedAttributes add spare columns to the schema that can be assigned to attributes at runtime.
@@ -252,25 +268,31 @@ type Trace struct {
 
 func attrToParquet(a *v1.KeyValue, p *Attribute, counter droppedAttrCounter) {
 	p.Key = a.Key
-	p.Value = nil
-	p.ValueBool = nil
-	p.ValueDouble = nil
-	p.ValueInt = nil
+	p.ValueType = 0
+	p.Value = p.Value[:0]
+	p.ValueInt = p.ValueInt[:0]
+	p.ValueDouble = p.ValueDouble[:0]
+	p.ValueBool = p.ValueBool[:0]
 	p.ValueDropped = ""
 
 	switch v := a.GetValue().Value.(type) {
 	case *v1.AnyValue_StringValue:
-		p.Value = &v.StringValue
+		p.Value = append(p.Value, v.StringValue)
+		p.ValueType = attrTypeString
 	case *v1.AnyValue_IntValue:
-		p.ValueInt = &v.IntValue
+		p.ValueInt = append(p.ValueInt, v.IntValue)
+		p.ValueType = attrTypeInt
 	case *v1.AnyValue_DoubleValue:
-		p.ValueDouble = &v.DoubleValue
+		p.ValueDouble = append(p.ValueDouble, v.DoubleValue)
+		p.ValueType = attrTypeDouble
 	case *v1.AnyValue_BoolValue:
-		p.ValueBool = &v.BoolValue
+		p.ValueBool = append(p.ValueBool, v.BoolValue)
+		p.ValueType = attrTypeBool
 	default:
 		jsonBytes := &bytes.Buffer{}
 		_ = jsonMarshaler.Marshal(jsonBytes, a.Value) // deliberately marshalling a.Value because of AnyValue logic
 		p.ValueDropped = jsonBytes.String()
+		p.ValueType = attrTypeNotSupported
 		counter.addDroppedAttr(1)
 	}
 }
@@ -529,34 +551,46 @@ func parquetToProtoAttrs(parquetAttrs []Attribute, counter droppedAttrCounter, i
 	var protoAttrs []*v1.KeyValue
 
 	for _, attr := range parquetAttrs {
-		protoVal := &v1.AnyValue{}
+		var protoVal v1.AnyValue
 
-		if attr.Value != nil {
-			protoVal.Value = &v1.AnyValue_StringValue{
-				StringValue: *attr.Value,
+		switch attr.ValueType {
+		case attrTypeString:
+			var v v1.AnyValue_StringValue
+			if len(attr.Value) > 0 {
+				v.StringValue = attr.Value[0]
 			}
-		} else if attr.ValueInt != nil {
-			protoVal.Value = &v1.AnyValue_IntValue{
-				IntValue: *attr.ValueInt,
+			protoVal.Value = &v
+		case attrTypeInt:
+			var v v1.AnyValue_IntValue
+			if len(attr.ValueInt) > 0 {
+				v.IntValue = attr.ValueInt[0]
 			}
-		} else if attr.ValueDouble != nil {
-			protoVal.Value = &v1.AnyValue_DoubleValue{
-				DoubleValue: *attr.ValueDouble,
+			protoVal.Value = &v
+		case attrTypeDouble:
+			var v v1.AnyValue_DoubleValue
+			if len(attr.ValueDouble) > 0 {
+				v.DoubleValue = attr.ValueDouble[0]
 			}
-		} else if attr.ValueBool != nil {
-			protoVal.Value = &v1.AnyValue_BoolValue{
-				BoolValue: *attr.ValueBool,
+			protoVal.Value = &v
+		case attrTypeBool:
+			var v v1.AnyValue_BoolValue
+			if len(attr.ValueBool) > 0 {
+				v.BoolValue = attr.ValueBool[0]
 			}
-		} else if attr.ValueDropped != "" && includeDroppedAttr {
-			_ = jsonpb.Unmarshal(bytes.NewBufferString(attr.ValueDropped), protoVal)
+			protoVal.Value = &v
+		case attrTypeNotSupported:
+			if attr.ValueDropped == "" || !includeDroppedAttr {
+				continue
+			}
+			_ = jsonpb.Unmarshal(bytes.NewBufferString(attr.ValueDropped), &protoVal)
 			counter.addDroppedAttr(-1)
-		} else {
+		default:
 			continue
 		}
 
 		protoAttrs = append(protoAttrs, &v1.KeyValue{
 			Key:   attr.Key,
-			Value: protoVal,
+			Value: &protoVal,
 		})
 	}