… tempodb/encoding/vparquet3/dedicated_columns_test.go
… tempodb/encoding/vparquet3/schema_test.go
… vendor/github.com/google/go-cmp/cmp/cmpopts/equate.go
… vendor/github.com/google/go-cmp/cmp/cmpopts/ignore.go
… vendor/github.com/google/go-cmp/cmp/cmpopts/sort.go
… vendor/github.com/google/go-cmp/cmp/cmpopts/struct_filter.go
… vendor/github.com/google/go-cmp/cmp/cmpopts/xform.go
… vendor/modules.txt