**What this PR does**:

This PR adds support for storing arrays with homogeneous, primitive types in Tempo blocks

**Which issue(s) this PR fixes**:
Fixes grafana/tempo-squad#360

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`