@stoewer 2023-12-12
The array support is implemented by storing all attribute values as slices / repeatable values. This comes with some overhead. Here are the latest benchmark results:

```
                                                 │ bench-arrays-base2.txt │        bench-arrays-new2.txt         │
                                                 │         sec/op         │    sec/op     vs base                │
BackendBlockTraceQL/spanAttValNoMatch-16                      10.90m ± 1%    10.99m ± 1%   +0.83% (p=0.001 n=12)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                10.22m ± 1%    10.24m ± 0%        ~ (p=0.078 n=12)
BackendBlockTraceQL/resourceAttValNoMatch-16                  10.12m ± 0%    10.16m ± 0%   +0.34% (p=0.039 n=12)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16              29.28m ± 0%    30.32m ± 0%   +3.57% (p=0.000 n=12)
BackendBlockTraceQL/mixedValNoMatch-16                        218.6m ± 1%    213.7m ± 1%   -2.24% (p=0.002 n=12)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                  10.12m ± 0%    10.19m ± 1%   +0.69% (p=0.010 n=12)
BackendBlockTraceQL/mixedValMixedMatchOr-16                   223.9m ± 2%    228.9m ± 2%   +2.24% (p=0.033 n=12)
BackendBlockTraceQL/count-16                                  363.0m ± 1%    365.9m ± 1%   +0.81% (p=0.024 n=12)
BackendBlockTraceQL/struct-16                                 617.6m ± 4%    632.0m ± 2%        ~ (p=0.143 n=12)
BackendBlockTraceQL/||-16                                     98.60m ± 0%   112.96m ± 0%  +14.57% (p=0.000 n=12)
BackendBlockTraceQL/mixed-16                                  50.68m ± 1%    54.57m ± 0%   +7.68% (p=0.000 n=12)
geomean                                                       56.39m         57.93m        +2.73%

                                                 │ bench-arrays-base2.txt │        bench-arrays-new2.txt         │
                                                 │          B/s           │     B/s       vs base                │
BackendBlockTraceQL/spanAttValNoMatch-16                     116.2Mi ± 1%   115.3Mi ± 1%   -0.82% (p=0.001 n=12)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16               115.6Mi ± 1%   115.4Mi ± 0%        ~ (p=0.078 n=12)
BackendBlockTraceQL/resourceAttValNoMatch-16                 19.32Mi ± 0%   19.26Mi ± 0%   -0.32% (p=0.040 n=12)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16             460.0Mi ± 0%   444.1Mi ± 0%   -3.45% (p=0.000 n=12)
BackendBlockTraceQL/mixedValNoMatch-16                       7.024Mi ± 1%   7.186Mi ± 2%   +2.31% (p=0.002 n=12)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                 20.63Mi ± 0%   20.49Mi ± 1%   -0.67% (p=0.008 n=12)
BackendBlockTraceQL/mixedValMixedMatchOr-16                  42.34Mi ± 2%   41.41Mi ± 2%   -2.21% (p=0.030 n=12)
BackendBlockTraceQL/count-16                                 39.47Mi ± 1%   39.14Mi ± 1%   -0.83% (p=0.018 n=12)
BackendBlockTraceQL/struct-16                                5.112Mi ± 4%   4.997Mi ± 2%        ~ (p=0.147 n=12)
BackendBlockTraceQL/||-16                                    148.9Mi ± 0%   129.9Mi ± 0%  -12.74% (p=0.000 n=12)
BackendBlockTraceQL/mixed-16                                 264.0Mi ± 1%   226.9Mi ± 0%  -14.07% (p=0.000 n=12)
geomean                                                      50.80Mi        49.10Mi        -3.34%

                                                 │ bench-arrays-base2.txt │        bench-arrays-new2.txt         │
                                                 │        MB_io/op        │  MB_io/op    vs base                 │
BackendBlockTraceQL/spanAttValNoMatch-16                       1.329 ± 0%    1.329 ± 0%       ~ (p=1.000 n=12) ¹
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                 1.239 ± 0%    1.239 ± 0%       ~ (p=1.000 n=12) ¹
BackendBlockTraceQL/resourceAttValNoMatch-16                  205.1m ± 0%   205.2m ± 0%  +0.05% (p=0.000 n=12)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16               14.12 ± 0%    14.12 ± 0%       ~ (p=1.000 n=12) ¹
BackendBlockTraceQL/mixedValNoMatch-16                         1.609 ± 0%    1.609 ± 0%       ~ (p=1.000 n=12) ¹
BackendBlockTraceQL/mixedValMixedMatchAnd-16                  218.9m ± 0%   218.9m ± 0%       ~ (p=1.000 n=12) ¹
BackendBlockTraceQL/mixedValMixedMatchOr-16                    9.938 ± 0%    9.937 ± 0%  -0.01% (p=0.000 n=12)
BackendBlockTraceQL/count-16                                   15.02 ± 0%    15.01 ± 0%  -0.07% (p=0.000 n=12)
BackendBlockTraceQL/struct-16                                  3.309 ± 0%    3.311 ± 0%  +0.06% (p=0.000 n=12)
BackendBlockTraceQL/||-16                                      15.39 ± 0%    15.39 ± 0%       ~ (p=1.000 n=12) ¹
BackendBlockTraceQL/mixed-16                                   14.03 ± 0%    12.98 ± 0%  -7.48% (p=0.000 n=12)
geomean                                                        3.003         2.982       -0.70%
¹ all samples are equal

                                                 │ bench-arrays-base2.txt │        bench-arrays-new2.txt        │
                                                 │          B/op          │     B/op      vs base               │
BackendBlockTraceQL/spanAttValNoMatch-16                     7.969Mi ± 0%   8.426Mi ± 1%  +5.73% (p=0.000 n=12)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16               7.058Mi ± 1%   7.491Mi ± 0%  +6.13% (p=0.000 n=12)
BackendBlockTraceQL/resourceAttValNoMatch-16                 6.977Mi ± 0%   7.403Mi ± 0%  +6.10% (p=0.000 n=12)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16             8.455Mi ± 3%   8.821Mi ± 2%  +4.33% (p=0.000 n=12)
BackendBlockTraceQL/mixedValNoMatch-16                       8.599Mi ± 2%   8.960Mi ± 4%  +4.20% (p=0.000 n=12)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                 7.047Mi ± 0%   7.461Mi ± 0%  +5.88% (p=0.000 n=12)
BackendBlockTraceQL/mixedValMixedMatchOr-16                  7.660Mi ± 3%   7.953Mi ± 3%  +3.83% (p=0.010 n=12)
BackendBlockTraceQL/count-16                                 261.6Mi ± 1%   262.5Mi ± 1%       ~ (p=0.478 n=12)
BackendBlockTraceQL/struct-16                                27.31Mi ± 3%   27.41Mi ± 3%       ~ (p=0.843 n=12)
BackendBlockTraceQL/||-16                                    15.98Mi ± 1%   16.31Mi ± 1%  +2.11% (p=0.000 n=12)
BackendBlockTraceQL/mixed-16                                 7.265Mi ± 2%   7.802Mi ± 1%  +7.39% (p=0.000 n=12)
geomean                                                      12.61Mi        13.13Mi       +4.19%

                                                 │ bench-arrays-base2.txt │       bench-arrays-new2.txt        │
                                                 │       allocs/op        │  allocs/op   vs base               │
BackendBlockTraceQL/spanAttValNoMatch-16                      109.5k ± 0%   112.7k ± 0%  +2.91% (p=0.000 n=12)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                109.4k ± 0%   112.6k ± 0%  +2.91% (p=0.000 n=12)
BackendBlockTraceQL/resourceAttValNoMatch-16                  109.4k ± 0%   112.6k ± 0%  +2.91% (p=0.000 n=12)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16              116.2k ± 0%   119.3k ± 0%  +2.67% (p=0.000 n=12)
BackendBlockTraceQL/mixedValNoMatch-16                        111.8k ± 0%   115.0k ± 0%  +2.81% (p=0.000 n=12)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                  109.5k ± 0%   112.6k ± 0%  +2.90% (p=0.000 n=12)
BackendBlockTraceQL/mixedValMixedMatchOr-16                   110.8k ± 0%   113.9k ± 0%  +2.79% (p=0.000 n=12)
BackendBlockTraceQL/count-16                                  1.513M ± 0%   1.516M ± 0%  +0.20% (p=0.000 n=12)
BackendBlockTraceQL/struct-16                                 332.3k ± 0%   334.8k ± 0%  +0.76% (p=0.005 n=12)
BackendBlockTraceQL/||-16                                     181.3k ± 0%   184.3k ± 0%  +1.66% (p=0.000 n=12)
BackendBlockTraceQL/mixed-16                                  111.2k ± 0%   114.3k ± 0%  +2.77% (p=0.000 n=12)
geomean                                                       162.5k        166.3k       +2.29%
```